test: . Move isBlocked() and isBlocking() to Task.ts
